# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a management backend API server.",
        "title": "Management BE API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/players": {
            "get": {
                "description": "Get a list of all players with their basic statistics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "List all players",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/player.ListPlayersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new player with their details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Create a new player",
                "parameters": [
                    {
                        "description": "Player information",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player.CreatePlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/player.PlayerResponse"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/players/{id}": {
            "put": {
                "description": "Update an existing player's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Update a player",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Player information",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/player.UpdatePlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/player.PlayerResponse"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a player by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Delete a player",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/player.DeletePlayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "player.CreatePlayerRequest": {
            "type": "object",
            "required": [
                "department_id",
                "full_name",
                "jersey_number",
                "position"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-01"
                },
                "department_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "John Doe"
                },
                "height_cm": {
                    "type": "integer",
                    "maximum": 300,
                    "minimum": 50,
                    "example": 180
                },
                "jersey_number": {
                    "type": "integer",
                    "maximum": 999,
                    "minimum": 1,
                    "example": 10
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "+1234567890"
                },
                "position": {
                    "type": "string",
                    "example": "Forward"
                },
                "weight_kg": {
                    "type": "integer",
                    "maximum": 200,
                    "minimum": 30,
                    "example": 75
                }
            }
        },
        "player.DeletePlayerResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Player deleted successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "player.ListPlayersResponse": {
            "type": "object",
            "properties": {
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/player.PlayerListItem"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "player.PlayerBasicStats": {
            "type": "object",
            "properties": {
                "total_assists": {
                    "type": "integer",
                    "example": 3
                },
                "total_goals": {
                    "type": "integer",
                    "example": 5
                },
                "total_matches": {
                    "type": "integer",
                    "example": 10
                },
                "total_minutes_played": {
                    "type": "integer",
                    "example": 900
                }
            }
        },
        "player.PlayerListItem": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-01"
                },
                "department_id": {
                    "type": "integer",
                    "example": 1
                },
                "department_name": {
                    "type": "string",
                    "example": "Engineering"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "height_cm": {
                    "type": "integer",
                    "example": 180
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "jersey_number": {
                    "type": "integer",
                    "example": 10
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "position": {
                    "type": "string",
                    "example": "Forward"
                },
                "statistics": {
                    "$ref": "#/definitions/player.PlayerBasicStats"
                },
                "weight_kg": {
                    "type": "integer",
                    "example": 75
                }
            }
        },
        "player.PlayerResponse": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "department_id": {
                    "type": "integer"
                },
                "department_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "height_cm": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "weight_kg": {
                    "type": "integer"
                }
            }
        },
        "player.UpdatePlayerRequest": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-01"
                },
                "department_id": {
                    "type": "integer",
                    "example": 1
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "height_cm": {
                    "type": "integer",
                    "example": 180
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "jersey_number": {
                    "type": "integer",
                    "example": 10
                },
                "phone": {
                    "type": "string",
                    "example": "+1234567890"
                },
                "position": {
                    "type": "string",
                    "example": "Forward"
                },
                "weight_kg": {
                    "type": "integer",
                    "example": 75
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}